+++ GitUI/Resources/ChangeLog.md
@@ -2,6 +2,7 @@
 =========
 ### Version 2.48.03 (9 December 2014)
 * Fixed issue #2538: Fix crash happening when deleting a remote branch
+* Fixed issue #2498: VS Plugin use solution scope if no active document
 
 ### Version 2.48.02 (29 November 2014)
 * Updated msysgit to 1.9.4 20140929